pulled out schemas componenet + styles
